+++ comics2pdf/command.py
@@ -1,0 +1,3 @@
+def main():
+    from comics2pdf import main_func
+    main_func()
